# src/main/resources/public/products-api-specification.yml
openapi: 3.0.3
info:
  title: Products API
  contact:
    name: Product Team
    email: reachanteneh@gmail.com
  version: v0.1
servers:
  - url: http://localhost:8080/api/v1
    description: local dev environment
security:
  - BasicAuth: [ ]
tags:
  - name: products-api
paths:
  /product-catalogue:
    post:
      tags:
        - product-catalogue-api
      summary: Create Products
      operationId: createProducts
      requestBody:
        description: products
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCatalogue'
        required: true
      responses:
        201:
          description: Created
          content: { }
        401:
          description: Unauthorized
          content: { }
        403:
          description: Forbidden
          content: { }
        404:
          description: Not Found
          content: { }
  /products:
    get:
      tags:
        - products-api
      summary: Products, optionally filtered
      operationId: getProducts
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: page
          in: query
          description: |
            Page number of the requested page. Defaults to first page.
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: |
            Size of a page. Defaults to 20 items.
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          description: |
            Sorting criteria in the form of `<property-name>`.
            The property name, MUST be from the `Product` property names.
            For multiple sort values, use multiple `sort` request parameters.
            Using multiple sort values in one request parameter will not work.
          explode: true
          schema:
            type: array
            items:
              type: string
          examples:
            single:
              value: [ 'name,asc' ]
              summary: Singular sort
            multi:
              value: [ 'name,asc','created_at,asc' ]
              summary: Multi sort
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsPage'
        401:
          description: Unauthorized
          content: { }
        403:
          description: Forbidden
          content: { }
    post:
      tags:
        - products-api
      summary: Create a Product
      operationId: createProduct
      requestBody:
        description: product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        201:
          description: Created
          content: { }
        401:
          description: Unauthorized
          content: { }
        403:
          description: Forbidden
          content: { }
        404:
          description: Not Found
          content: { }
    delete:
      tags:
        - products-api
      summary: Delete all products which match the selection criteria (not available in live environment)
      operationId: deleteAllProducts
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        200:
          description: Ok
          content: { }
        401:
          description: Unauthorized
          content: { }
        403:
          description: Forbidden
          content: { }
  /products/{id}:
    get:
      tags:
        - products-api
      summary: Get product by id
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        401:
          description: Unauthorized
          content: { }
        403:
          description: Forbidden
          content: { }
components:
  securitySchemes:
    BasicAuth:
      description: >
        Basic authentication
      type: http
      scheme: basic
  schemas:
    ProductsPage:
      title: ProductsPage
      type: object
      properties:
        items:
          title: Array of Products
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          $ref: './common-model.yml#/components/schemas/Page'
    Product:
      title: Product
      type: object
      required: [ name, price, contain_articles ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        identifier:
          type: string
        price:
          type: number
          format: double
          default: 0.0
        available_quantity:
          type: integer
          format: int32
        contain_articles:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ContainArticle'
    ContainArticle:
      title: ContainArticle
      required: [ art_id, amount_of ]
      type: object
      properties:
        art_id:
          type: string
        amount_of:
          type: integer
          format: int32
    ProductCatalogue:
      title: Products
      type: object
      required: [ products ]
      properties:
        products:
          title: Array of Products
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Product'